+++ src/index.ts
@@ -287,6 +287,14 @@ Hooks.on('ready', async () => {
             requiresReload: true,
             type: Boolean,
         } as any);
+        gameInstance.settings.register('pf2e-kingmaker-tools', 'reduceDCToBuildLumberStructures', {
+            name: 'Reduce DC to Build Lumber Structures',
+            scope: 'world',
+            config: false,
+            default: false,
+            requiresReload: true,
+            type: Boolean,
+        } as any);
         gameInstance.settings.register('pf2e-kingmaker-tools', 'kingdomAlwaysAddHalfLevel', {
             name: 'Always add half Level to Skill',
             hint: 'If enabled, always adds half of the kingdom\'s level to a skill, even if it is untrained',