+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.27.0] - 2024-02-02
+
+### Added
+
+* Setting to automate Ekundayo's -2 to lumber buildings reduction
+* Also post description of activity to chat to display resource buttons
+* Add buttons to gain commodities in Purchase Commodities result
+
 ### Fixed
 
 * Make changing degree of success work again on kingdom activity results